feat: 구간관련 테스트 RestAssured로 이전
