+++ src/test/java/wooteco/subway/section/web/SectionApiControllerTest.java
@@ -1,398 +1,398 @@
 package wooteco.subway.section.web;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.MediaType;
-import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.http.HttpStatus;
 import org.springframework.test.web.servlet.ResultActions;
-import org.springframework.transaction.annotation.Transactional;
-import wooteco.subway.ApiControllerTest;
+import wooteco.subway.AcceptanceTest;
 import wooteco.subway.domain.Line;
-import wooteco.subway.domain.Section;
 import wooteco.subway.domain.Sections;
 import wooteco.subway.domain.Station;
-import wooteco.subway.line.LineService;
-import wooteco.subway.section.dao.SectionDao;
-import wooteco.subway.station.dao.StationDao;
+import wooteco.subway.line.web.LineRequest;
+import wooteco.subway.station.StationRequest;
+import wooteco.subway.station.StationResponse;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
-@Transactional
-@DisplayName("구간관련 테스트")
-class SectionApiControllerTest extends ApiControllerTest {
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-    private static final String UP_STATION_NAME = "잠실역";
-    private static final String DOWN_STATION_NAME = "잠실새내역";
-    private static final int ORIGINAL_DISTANCE = 10;
+@DisplayName("[API] 구간관련 테스트")
+class SectionApiControllerTest extends AcceptanceTest {
+    private StationRequest 잠실역;
+    private StationRequest 잠실새내역;
+    private StationRequest 강남역;
+    private StationRequest 동탄역;
+    private StationRequest 수서역;
+    private int ORIGINAL_DISTANCE;
 
-    @Autowired
-    private MockMvc mockMvc;
-    @Autowired
-    private ObjectMapper objectMapper;
-    @Autowired
-    private StationDao stationDao;
-    @Autowired
-    private LineService lineService;
-    @Autowired
-    private SectionDao sectionDao;
+    @BeforeEach
+    void setUpSectionTest() {
+        잠실역 = new StationRequest("잠실역");
+        잠실새내역 = new StationRequest("잠실새내역");
+        강남역 = new StationRequest("강남역");
+        동탄역 = new StationRequest("동탄역");
+        수서역 = new StationRequest("수서역");
+        ORIGINAL_DISTANCE = 10;
+    }
 
     @Test
     @DisplayName("구간 등록 - 성공(상행종점 등록)")
-    void create_성공_상행종점추가() throws Exception {
-        final Station 잠실역 = 상행역();
-        final Station 잠실새내역 = 하행역();
-        Line line = 노선_생성(잠실역, 잠실새내역);
+    void create_성공_상행종점추가() {
+        Long 잠실역_id = postStationAndGetId(잠실역);
+        Long 잠실새내_id = postStationAndGetId(잠실새내역);
+        final LineRequest 이호선 =
+                new LineRequest("2호선", "bg-green-600", 잠실역_id, 잠실새내_id, ORIGINAL_DISTANCE);
+        Long lineId = postLineAndGetId(이호선);
 
-        Station 강남역 = stationDao.create(Station.create("강남역"));
-        SectionRequest sectionRequest = new SectionRequest(강남역.getId(), 잠실역.getId(), 4);
+        Long 강남역_id = postStationAndGetId(강남역);
+        SectionRequest 강남_잠실 = new SectionRequest(강남역_id, 잠실역_id, 4);
 
         // when
-        ResultActions result = 구간_추가(sectionRequest, line.getId());
-        final List<Section> sections = sectionDao.findAllByLineId(line.getId()).sections();
+        ExtractableResponse<Response> result = postSection(강남_잠실, lineId);
+        ExtractableResponse<Response> lineResult = get("/lines/" + lineId);
+        List<String> stationsResult = lineResult.jsonPath().getList("stations", StationResponse.class)
+                .stream()
+                .map(StationResponse::getName)
+                .collect(Collectors.toList());
+
 
         // then
-        result.andDo(print())
-                .andExpect(status().isCreated())
-                .andExpect(header().exists("Location"));
-        assertThat(sections).hasSize(2);
+        assertThat(result.statusCode()).isEqualTo(CREATED);
+        assertThat(result.header("Location")).isNotNull();
+        assertThat(stationsResult).hasSize(3);
+        assertThat(stationsResult).containsExactly("강남역", "잠실역", "잠실새내역");
     }
 
     @Test
     @DisplayName("구간 등록 - 성공(하행종점 등록)")
-    void create_성공_하행종점추가() throws Exception {
+    void create_성공_하행종점추가() {
         // given
-        final Station 잠실역 = 상행역();
-        final Station 잠실새내역 = 하행역();
-        Station 강남역 = stationDao.create(Station.create("강남역"));
+        Long 잠실역_id = postStationAndGetId(잠실역);
+        Long 잠실새내_id = postStationAndGetId(잠실새내역);
+        final LineRequest 이호선 =
+                new LineRequest("2호선", "bg-green-600", 잠실역_id, 잠실새내_id, ORIGINAL_DISTANCE);
+        Long lineId = postLineAndGetId(이호선);
 
-        Line line = 노선_생성(잠실역, 잠실새내역);
-
-        SectionRequest sectionRequest = new SectionRequest(잠실새내역.getId(), 강남역.getId(), 4);
+        Long 강남역_id = postStationAndGetId(강남역);
+        SectionRequest 잠실새내_강남 = new SectionRequest(잠실새내_id, 강남역_id, 4);
 
         // when
-        ResultActions result = 구간_추가(sectionRequest, line.getId());
-        final List<Section> sections = sectionDao.findAllByLineId(line.getId()).sections();
+        ExtractableResponse<Response> result = postSection(잠실새내_강남, lineId);
+        ExtractableResponse<Response> lineResult = get("/lines/" + lineId);
+        List<String> stationsResult = lineResult.jsonPath().getList("stations", StationResponse.class)
+                .stream()
+                .map(StationResponse::getName)
+                .collect(Collectors.toList());
 
         // then
-        result.andDo(print())
-                .andExpect(status().isCreated())
-                .andExpect(header().exists("Location"));
-
-        assertThat(sections).hasSize(2);
+        assertThat(result.statusCode()).isEqualTo(CREATED);
+        assertThat(result.header("Location")).isNotNull();
+        assertThat(stationsResult).hasSize(3);
+        assertThat(stationsResult).containsExactly("잠실역", "잠실새내역", "강남역");
     }
 
     @Test
     @DisplayName("구간 등록 - 성공(상행기준 중간구간 구간 등록)")
-    void create_성공_중간역상행기준() throws Exception {
+    void create_성공_중간역상행기준() {
         // given
-        final Station 잠실역 = 상행역();
-        final Station 잠실새내역 = 하행역();
-        Station 강남역 = stationDao.create(Station.create("강남역"));
-
-        Line line = 노선_생성(잠실역, 잠실새내역);
-
-        SectionRequest sectionRequest = new SectionRequest(잠실역.getId(), 강남역.getId(), 4);
+        Long 잠실역_id = postStationAndGetId(잠실역);
+        Long 잠실새내_id = postStationAndGetId(잠실새내역);
+        final LineRequest 이호선 =
+                new LineRequest("2호선", "bg-green-600", 잠실역_id, 잠실새내_id, ORIGINAL_DISTANCE);
+        Long lineId = postLineAndGetId(이호선);
+        Long 강남역_id = postStationAndGetId(강남역);
+        SectionRequest 잠실_강남 = new SectionRequest(잠실역_id, 강남역_id, 4);
 
         // when
-        ResultActions result = 구간_추가(sectionRequest, line.getId());
-        final List<Section> sections = sectionDao.findAllByLineId(line.getId()).sections();
+        ExtractableResponse<Response> result = postSection(잠실_강남, lineId);
+        ExtractableResponse<Response> lineResult = get("/lines/" + lineId);
+        List<String> stationsResult = lineResult.jsonPath().getList("stations", StationResponse.class)
+                .stream()
+                .map(StationResponse::getName)
+                .collect(Collectors.toList());
 
         // then
-        result.andDo(print())
-                .andExpect(status().isCreated())
-                .andExpect(header().exists("Location"));
-
-        assertThat(sections).hasSize(2);
+        assertThat(result.statusCode()).isEqualTo(CREATED);
+        assertThat(result.header("Location")).isNotNull();
+        assertThat(stationsResult).hasSize(3);
+        assertThat(stationsResult).containsExactly("잠실역", "강남역", "잠실새내역");
     }
 
     @Test
     @DisplayName("구간 등록 - 성공(하행기준 중간구간 구간 등록)")
-    void create_성공_중간하행기준() throws Exception {
+    void create_성공_중간하행기준() {
         // given
-        final Station 잠실역 = 상행역();
-        final Station 잠실새내역 = 하행역();
-        Station 강남역 = stationDao.create(Station.create("강남역"));
+        Long 잠실역_id = postStationAndGetId(잠실역);
+        Long 잠실새내_id = postStationAndGetId(잠실새내역);
+        final LineRequest 이호선 =
+                new LineRequest("2호선", "bg-green-600", 잠실역_id, 잠실새내_id, ORIGINAL_DISTANCE);
+        Long lineId = postLineAndGetId(이호선);
 
-        Line line = 노선_생성(잠실역, 잠실새내역);
-
-        SectionRequest 강남_잠실새내 = new SectionRequest(강남역.getId(), 잠실새내역.getId(), 4);
+        Long 강남역_id = postStationAndGetId(강남역);
+        SectionRequest 강남_잠실새내 = new SectionRequest(강남역_id, 잠실새내_id, 4);
 
         // when
-        ResultActions result = 구간_추가(강남_잠실새내, line.getId());
-        final List<Section> sections = sectionDao.findAllByLineId(line.getId()).sections();
+        ExtractableResponse<Response> result = postSection(강남_잠실새내, lineId);
+        ExtractableResponse<Response> lineResult = get("/lines/" + lineId);
+        List<String> stationsResult = lineResult.jsonPath().getList("stations", StationResponse.class)
+                .stream()
+                .map(StationResponse::getName)
+                .collect(Collectors.toList());
 
         // then
-        result.andDo(print())
-                .andExpect(status().isCreated())
-                .andExpect(header().exists("Location"));
-
-        assertThat(sections).hasSize(2);
+        assertThat(result.statusCode()).isEqualTo(CREATED);
+        assertThat(result.header("Location")).isNotNull();
+        assertThat(stationsResult).hasSize(3);
+        assertThat(stationsResult).containsExactly("잠실역", "강남역", "잠실새내역");
     }
 
+    //todo: 실패 왜 안되지!!??!?!?! 이거 통과 안되는지 왜 의문
     @Test
     @DisplayName("구간 등록 - 성공(중간 구간 등록 a-b-c-d --(b-k)--> a-b-k-c-d)")
-    void create_성공_중간앞() throws Exception {
+    void create_성공_중간앞() {
         // given
-        final Station 잠실역 = 상행역();
-        final Station 잠실새내역 = 하행역();
-        Station 강남역 = stationDao.create(Station.create("강남역"));
-        Station 동탄역 = stationDao.create(Station.create("동탄역"));
-        Station 수서역 = stationDao.create(Station.create("수서역"));
+        Long 잠실역_id = postStationAndGetId(잠실역);
+        Long 잠실새내_id = postStationAndGetId(잠실새내역);
+        Long 강남역_id = postStationAndGetId(강남역);
+        Long 수서역_id = postStationAndGetId(수서역);
+        Long 동탄역_id = postStationAndGetId(동탄역);
+        final LineRequest 이호선 =
+                new LineRequest("2호선", "bg-green-600", 잠실역_id, 잠실새내_id, ORIGINAL_DISTANCE);
+        Long lineId = postLineAndGetId(이호선);
 
-        Line line = 노선_생성(잠실역, 잠실새내역);
-
-        SectionRequest 강남_잠실 = new SectionRequest(강남역.getId(), 잠실역.getId(), 4);
-        SectionRequest 잠실새내_동탄 = new SectionRequest(잠실새내역.getId(), 동탄역.getId(), 4);
-
-        구간_추가(강남_잠실, line.getId());
-        구간_추가(잠실새내_동탄, line.getId());
+        postSection(new SectionRequest(강남역_id, 잠실역_id, 4), lineId);
+        postSection(new SectionRequest(잠실새내_id, 동탄역_id, 4), lineId);
 
         // when
-        SectionRequest 잠실_수서 = new SectionRequest(잠실역.getId(), 수서역.getId(), 2);
-        ResultActions result = 구간_추가(잠실_수서, line.getId());
-        final List<Section> sections = sectionDao.findAllByLineId(line.getId()).sections();
+        ExtractableResponse<Response> result = postSection(new SectionRequest(잠실역_id, 수서역_id, 2), lineId);
+        ExtractableResponse<Response> lineResult = get("/lines/" + lineId);
+        List<String> stationsResult = lineResult.jsonPath().getList("stations", StationResponse.class)
+                .stream()
+                .map(StationResponse::getName)
+                .collect(Collectors.toList());
 
         // then
-        result.andDo(print())
-                .andExpect(status().isCreated())
-                .andExpect(header().exists("Location"));
-
-        assertThat(sections).hasSize(4);
+        assertThat(result.statusCode()).isEqualTo(CREATED);
+        assertThat(result.header("Location")).isNotNull();
+        assertThat(stationsResult).hasSize(5);
+        assertThat(stationsResult).containsExactly("강남역", "잠실역", "수서역", "잠실새내역", "동탄역");
     }
 
+    //todo: 해야함
     @Test
     @DisplayName("구간 등록 - 성공(중간 구간 등록 a-b-c-d --(k-c)--> a-b-k-c-d)")
-    void create_성공_중간뒤() throws Exception {
+    void create_성공_중간뒤() {
         // given
-        final Station 잠실역 = 상행역();
-        final Station 잠실새내역 = 하행역();
-        Station 강남역 = stationDao.create(Station.create("강남역"));
-        Station 동탄역 = stationDao.create(Station.create("동탄역"));
-        Station 수서역 = stationDao.create(Station.create("수서역"));
+        Long 잠실역_id = postStationAndGetId(잠실역);
+        Long 잠실새내_id = postStationAndGetId(잠실새내역);
+        Long 강남역_id = postStationAndGetId(강남역);
+        Long 수서역_id = postStationAndGetId(수서역);
+        Long 동탄역_id = postStationAndGetId(동탄역);
 
-        Line line = 노선_생성(잠실역, 잠실새내역);
+        final LineRequest 이호선 =
+                new LineRequest("2호선", "bg-green-600", 잠실역_id, 잠실새내_id, ORIGINAL_DISTANCE);
+        Long lineId = postLineAndGetId(이호선);
 
-        SectionRequest 강남_잠실 = new SectionRequest(강남역.getId(), 잠실역.getId(), 4);
-        SectionRequest 잠실새내_동탄 = new SectionRequest(잠실새내역.getId(), 동탄역.getId(), 4);
+        postSection(new SectionRequest(강남역_id, 잠실역_id, 4), lineId);
+        postSection(new SectionRequest(잠실새내_id, 동탄역_id, 4), lineId);
 
-        구간_추가(강남_잠실, line.getId());
-        구간_추가(잠실새내_동탄, line.getId());
 
-        // when
-        SectionRequest 수서_잠실새내 = new SectionRequest(수서역.getId(), 잠실새내역.getId(), 2);
-        ResultActions result = 구간_추가(수서_잠실새내, line.getId());
-        final List<Section> sections = sectionDao.findAllByLineId(line.getId()).sections();
+        ExtractableResponse<Response> result = postSection(new SectionRequest(수서역_id, 잠실새내_id, 2), lineId);
+        ExtractableResponse<Response> lineResult = get("/lines/" + lineId);
+        List<String> stationsResult = lineResult.jsonPath().getList("stations", StationResponse.class)
+                .stream()
+                .map(StationResponse::getName)
+                .collect(Collectors.toList());
 
         // then
-        result.andDo(print())
-                .andExpect(status().isCreated())
-                .andExpect(header().exists("Location"));
-
-        assertThat(sections).hasSize(4);
+        assertThat(result.statusCode()).isEqualTo(CREATED);
+        assertThat(result.header("Location")).isNotNull();
+        assertThat(stationsResult).hasSize(5);
+        assertThat(stationsResult).containsExactly("강남역", "잠실역", "수서역", "잠실새내역", "동탄역");
     }
 
     @Test
     @DisplayName("구간 등록 - 실패(새로 추가할 거리가 기존 거리보다 같거나 큰 경우)")
-    void create_실패_거리큼() throws Exception {
+    void create_실패_거리큼() {
         // given
-        final Station 잠실역 = 상행역();
-        final Station 잠실새내역 = 하행역();
-        Station 강남역 = stationDao.create(Station.create("강남역"));
-        final Line line = 노선_생성(잠실역, 잠실새내역);
+        Long 잠실역_id = postStationAndGetId(잠실역);
+        Long 잠실새내_id = postStationAndGetId(잠실새내역);
+        Long 강남역_id = postStationAndGetId(강남역);
+        final LineRequest 이호선 =
+                new LineRequest("2호선", "bg-green-600", 잠실역_id, 잠실새내_id, ORIGINAL_DISTANCE);
+        Long lineId = postLineAndGetId(이호선);
 
-        final SectionRequest 강남_잠실새내 =
-                new SectionRequest(강남역.getId(), 잠실새내역.getId(), ORIGINAL_DISTANCE);
-
-        // when
-        ResultActions result = 구간_추가(강남_잠실새내, line.getId());
+        ExtractableResponse<Response> result = postSection(new SectionRequest(강남역_id, 잠실새내_id, ORIGINAL_DISTANCE), lineId);
 
         // then
-        result.andDo(print())
-                .andExpect(status().isBadRequest())
-                .andExpect(content().string("새로 추가할 거리가 기존 거리보다 같거나 큽니다."));
+        assertThat(result.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
+        assertThat(result.body().asString()).isEqualTo("새로 추가할 거리가 기존 거리보다 같거나 큽니다.");
     }
 
     @Test
     @DisplayName("구간 등록 - 실패(존재하지 않는 역을 등록할 경우)")
-    void create_실패_역없음() throws Exception {
+    void create_실패_역없음() {
         // given
-        final Station 잠실역 = 상행역();
-        final Station 잠실새내역 = 하행역();
-        final Line line = 노선_생성(잠실역, 잠실새내역);
+        Long 잠실역_id = postStationAndGetId(잠실역);
+        Long 잠실새내_id = postStationAndGetId(잠실새내역);
+        final LineRequest 이호선 =
+                new LineRequest("2호선", "bg-green-600", 잠실역_id, 잠실새내_id, ORIGINAL_DISTANCE);
+        Long lineId = postLineAndGetId(이호선);
 
-        final SectionRequest sectionRequest =
-                new SectionRequest(Long.MAX_VALUE, 잠실새내역.getId(), ORIGINAL_DISTANCE);
+        final SectionRequest 이상한구간 = new SectionRequest(Long.MAX_VALUE, 잠실새내_id, ORIGINAL_DISTANCE);
 
         // when
-        final ResultActions result = 구간_추가(sectionRequest, line.getId());
+        ExtractableResponse<Response> result = postSection(이상한구간, lineId);
 
         //then
-        result.andDo(print())
-                .andExpect(status().isNotFound());
+        assertThat(result.statusCode()).isEqualTo(HttpStatus.NOT_FOUND.value());
     }
 
     @Test
     @DisplayName("구간 등록 - 실패(자연수가 아닌 거리를 등록할 경우)")
-    void create_실패_거리가자연수아님() throws Exception {
+    void create_실패_거리가자연수아님() {
         // given
-        final Station 잠실역 = 상행역();
-        final Station 잠실새내역 = 하행역();
-        final Station 강남역 = stationDao.create(Station.create("강남역"));
-        final Line line = 노선_생성(잠실역, 잠실새내역);
+        Long 잠실역_id = postStationAndGetId(잠실역);
+        Long 잠실새내_id = postStationAndGetId(잠실새내역);
+        Long 강남역_id = postStationAndGetId(강남역);
+        final LineRequest 이호선 =
+                new LineRequest("2호선", "bg-green-600", 잠실역_id, 잠실새내_id, ORIGINAL_DISTANCE);
+        Long lineId = postLineAndGetId(이호선);
 
-        final SectionRequest 강남_잠실새내 =
-                new SectionRequest(강남역.getId(), 잠실새내역.getId(), 0);
+        final SectionRequest 강남_잠실새내 = new SectionRequest(강남역_id, 잠실새내_id, 0);
 
         // when
-        final ResultActions result = 구간_추가(강남_잠실새내, line.getId());
+        ExtractableResponse<Response> result = postSection(강남_잠실새내, lineId);
 
         // then
-        result.andDo(print())
-                .andExpect(status().isBadRequest())
-                .andExpect(content().string("자연수가 아닌 거리를 등록하셨습니다."));
+        assertThat(result.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
+        assertThat(result.body().asString()).isEqualTo("자연수가 아닌 거리를 등록하셨습니다.");
     }
-
 
     @Test
     @DisplayName("구간 등록 - 실패(의미상 중복된 섹션을 등록할 경우)")
-    void create_실패_의미상중복구간() throws Exception {
-        //given
-        final Station 잠실역 = 상행역();
-        final Station 잠실새내역 = 하행역();
-        final Line line = 노선_생성(잠실역, 잠실새내역);
-        final SectionRequest 잠실새내_잠실 = new SectionRequest(잠실새내역.getId(), 잠실역.getId(), 3);
+    void create_실패_의미상중복구간() {
+        // given
+        Long 잠실역_id = postStationAndGetId(잠실역);
+        Long 잠실새내_id = postStationAndGetId(잠실새내역);
+        final LineRequest 이호선 =
+                new LineRequest("2호선", "bg-green-600", 잠실역_id, 잠실새내_id, ORIGINAL_DISTANCE);
+        Long lineId = postLineAndGetId(이호선);
+        final SectionRequest 잠실새내_잠실 = new SectionRequest(잠실새내_id, 잠실역_id, 3);
 
         //when
-        ResultActions result = 구간_추가(잠실새내_잠실, line.getId());
+        ExtractableResponse<Response> result = postSection(잠실새내_잠실, lineId);
 
         //then
-        result.andDo(print())
-                .andExpect(status().isBadRequest())
-                .andExpect(content().string("중복된 구간입니다."));
+        assertThat(result.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
+        assertThat(result.body().asString()).isEqualTo("중복된 구간입니다.");
     }
 
     @Test
     @DisplayName("구간 등록 - 실패(새로운 섹션의 두 역이 모두 같은 노선에 포함된 경우)")
-    void create_실패_사이클도는구간() throws Exception {
+    void create_실패_사이클도는구간() {
         // given
-        final Station 잠실역 = 상행역();
-        final Station 잠실새내역 = 하행역();
-        final Station 강남역 = stationDao.create(Station.create("강남역"));
-        final Line line = 노선_생성(잠실역, 강남역);
+        Long 잠실역_id = postStationAndGetId(잠실역);
+        Long 잠실새내_id = postStationAndGetId(잠실새내역);
+        Long 강남역_id = postStationAndGetId(강남역);
+        final LineRequest 이호선 =
+                new LineRequest("2호선", "bg-green-600", 잠실역_id, 강남역_id, ORIGINAL_DISTANCE);
+        Long lineId = postLineAndGetId(이호선);
 
-        final SectionRequest 강남_잠실새내 = new SectionRequest(강남역.getId(), 잠실새내역.getId(), 3);
-        final SectionRequest 잠실_잠실새내 = new SectionRequest(잠실역.getId(), 잠실새내역.getId(), 3);
+        final SectionRequest 강남_잠실새내 = new SectionRequest(강남역_id, 잠실새내_id, 3);
+        final SectionRequest 잠실_잠실새내 = new SectionRequest(잠실역_id, 잠실새내_id, 3);
 
-        구간_추가(강남_잠실새내, line.getId());
+        postSection(강남_잠실새내, lineId);
 
         // when
-        final ResultActions result = 구간_추가(잠실_잠실새내, line.getId());
+        ExtractableResponse<Response> result = postSection(잠실_잠실새내, lineId);
 
         // then
-        result.andDo(print())
-                .andExpect(status().isBadRequest())
-                .andExpect(content().string("사이클이 생기는 구간입니다."));
+        assertThat(result.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
+        assertThat(result.body().asString()).isEqualTo("사이클이 생기는 구간입니다.");
     }
 
+    // todo: 이상
     @Test
     @DisplayName("구간 제거 - 성공")
-    void delete_성공() throws Exception {
+    void delete_성공() {
         // given
-        final Station 잠실역 = 상행역();
-        final Station 잠실새내역 = 하행역();
-        Line line = 노선_생성(잠실역, 잠실새내역);
-
-        Station 강남역 = stationDao.create(Station.create("강남역"));
+        Long 잠실역_id = postStationAndGetId(잠실역);
+        Long 잠실새내_id = postStationAndGetId(잠실새내역);
+        Long 강남역_id = postStationAndGetId(강남역);
+        final LineRequest 이호선 =
+                new LineRequest("2호선", "bg-green-600", 잠실역_id, 잠실새내_id, ORIGINAL_DISTANCE);
+        Long lineId = postLineAndGetId(이호선);
 
         SectionRequest 강남_잠실 =
-                new SectionRequest(강남역.getId(), 잠실역.getId(), 4);
+                new SectionRequest(강남역_id, 잠실역_id, 4);
+        postSection(강남_잠실, lineId);
 
-        구간_추가(강남_잠실, line.getId());
         // when
-        ResultActions result = mockMvc.perform(
-                delete("/lines/" + line.getId() + "/sections?stationId=" + 잠실역.getId()));
-        final Sections sections = sectionDao.findAllByLineId(line.getId());
+        ExtractableResponse<Response> deleteResponse = delete("/lines/" + lineId + "/sections?stationId=" + 잠실역_id);
+        ExtractableResponse<Response> result = get("/lines/" + lineId);
+        List<StationResponse> stations = result.body().jsonPath().getList("stations", StationResponse.class);
 
         // then
-        result.andDo(print())
-                .andExpect(status().isNoContent());
-
-        assertThat(sections.sections().size()).isEqualTo(1);
-        assertThat(sections.sections().get(0).getUpStation().getName()).isEqualTo("강남역");
-        assertThat(sections.sections().get(0).getDownStation().getName()).isEqualTo("잠실새내역");
-        assertThat(sections.sections().get(0).getDistance()).isEqualTo(14);
+        assertThat(deleteResponse.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
+        assertThat(stations).hasSize(2);
+        assertThat(stations.get(0).getName()).isEqualTo("강남역");
+        assertThat(stations.get(1).getName()).isEqualTo("잠실새내역");
+        // todo: 거리?!
+        //  assertThat(result.body().jsonPath().getInt("distance")).isEqualTo(14);
     }
 
     @Test
     @DisplayName("구간 제거 - 실패(노선이 존재하지 않을 시)")
-    public void delete_실패_노선없음() throws Exception {
+    public void delete_실패_노선없음()  {
         //given & when
-        final ResultActions result = mockMvc.perform(
-                delete("/lines/" + Long.MAX_VALUE + "/sections?stationId=1"));
-        //then
+        ExtractableResponse<Response> result = delete("/lines/" + Long.MAX_VALUE + "/sections?stationId=1");
 
-        result.andExpect(status().isNotFound());
+        //then
+        assertThat(result.statusCode()).isEqualTo(HttpStatus.NOT_FOUND.value());
     }
 
     @Test
     @DisplayName("구간 제거 - 실패(역이 해당 노선에 등록되어 있지 않을 시)")
-    public void delete_실패_역이노선에없음() throws Exception {
-        //given
-        final Station 잠실역 = 상행역();
-        final Station 잠실새내역 = 하행역();
-        Line line = 노선_생성(잠실역, 잠실새내역);
+    public void delete_실패_역이노선에없음() {
+        // given
+        Long 잠실역_id = postStationAndGetId(잠실역);
+        Long 잠실새내_id = postStationAndGetId(잠실새내역);
+
+        final LineRequest 이호선 =
+                new LineRequest("2호선", "bg-green-600", 잠실역_id, 잠실새내_id, ORIGINAL_DISTANCE);
+        Long lineId = postLineAndGetId(이호선);
 
         //when
-        final ResultActions result = mockMvc.perform(
-                delete("/lines/" + line.getId() + "/sections?stationId=" + Long.MAX_VALUE)
-        );
+        ExtractableResponse<Response> result = delete("/lines/" + lineId + "/sections?stationId=" + Long.MAX_VALUE);
 
         //then
-        result.andDo(print())
-                .andExpect(status().isNotFound());
+        assertThat(result.statusCode()).isEqualTo(HttpStatus.NOT_FOUND.value());
     }
 
     @Test
     @DisplayName("구간 제거 - 실패(노선에 구간이 하나밖에 존재하지 않을 시)")
-    public void delete_실패_마지막남은구간() throws Exception {
-        //given
-        final Station 잠실역 = 상행역();
-        final Station 잠실새내역 = 하행역();
-        Line line = 노선_생성(잠실역, 잠실새내역);
+    public void delete_실패_마지막남은구간() {
+        // given
+        Long 잠실역_id = postStationAndGetId(잠실역);
+        Long 잠실새내_id = postStationAndGetId(잠실새내역);
+        final LineRequest 이호선 =
+                new LineRequest("2호선", "bg-green-600", 잠실역_id, 잠실새내_id, ORIGINAL_DISTANCE);
+        Long lineId = postLineAndGetId(이호선);
 
         //when
-        final ResultActions result = mockMvc.perform(
-                delete("/lines/" + line.getId() + "/sections?stationId=" + 잠실새내역.getId())
-        );
+        ExtractableResponse<Response> result = delete("/lines/" + lineId + "/sections?stationId=" + 잠실새내_id);
 
         //then
-        result.andExpect(status().isBadRequest());
+        assertThat(result.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
     }
 
-    private ResultActions 구간_추가(SectionRequest sectionRequest, Long lineId) throws Exception {
-        return mockMvc.perform(post("/lines/" + lineId + "/sections")
-                .content(objectMapper.writeValueAsString(sectionRequest))
-                .accept(MediaType.APPLICATION_JSON_VALUE)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-        );
-    }
-
-    private Line 노선_생성(Station upStation, Station downStation) {
-        return lineService.create("1호선", "bg-blue-300", upStation, downStation, ORIGINAL_DISTANCE);
-    }
-
-    private Station 하행역() {
-        return stationDao.create(Station.create(DOWN_STATION_NAME));
-    }
-
-    private Station 상행역() {
-        return stationDao.create(Station.create(UP_STATION_NAME));
-    }
 }
